+++ client/src/execute.rs
@@ -12,33 +12,22 @@
 //! - Refund Bitcoin from HTLC after timeout expiry (buyer only)
 
 use std::path::PathBuf;
-use std::time::{Duration, Instant};
 
-use anchor_client::solana_sdk::signature::read_keypair_file;
 use anyhow::{Context, Result, anyhow};
 use bitcoin::{Amount, PublicKey};
 use bitcoincore_rpc::Auth;
 use btc_htlc::{Contract as BtcContract, HtlcParams};
-use ethers::types::{H256, U256};
 use sha2::{Digest, Sha256};
-use tokio::time::sleep;
 use tracing::{debug, info, instrument};
+use utils::{DEFAULT_SECRETS_DIR, DEFAULT_SECRETS_FILE};
 
 use crate::btc::BtcClient;
-use crate::eth::EthClient;
-use crate::sol::SolClient;
 use crate::types::{
     CancelCommitArgs, CancelResult, Chain, ClaimBtcArgs, ClaimBtcResult, CommitForMintArgs,
     CommitResult, LockBtcArgs, LockBtcResult, MintResult, MintWithSecretArgs, RefundBtcArgs,
     RefundBtcResult,
 };
-use crate::utils;
-
-/// Maximum time to wait for mint availability.
-const MINT_AVAILABILITY_TIMEOUT: Duration = Duration::from_secs(120);
-
-/// Interval between mint availability checks.
-const MINT_CHECK_INTERVAL: Duration = Duration::from_secs(5);
+use crate::{eth, sol, utils};
 
 /// Locks Bitcoin in an HTLC contract.
 ///
@@ -94,9 +83,9 @@ pub fn lock_bitcoin(args: LockBtcArgs) -> Result<LockBtcResult> {
 
     debug!(secret = %hex::encode(secret_bytes), "Generated secret");
 
-    let secret_file = args.secret_output_file.unwrap_or_else(|| {
-        PathBuf::from(utils::DEFAULT_SECRETS_DIR).join(utils::DEFAULT_SECRETS_FILE)
-    });
+    let secret_file = args
+        .secret_output_file
+        .unwrap_or_else(|| PathBuf::from(DEFAULT_SECRETS_DIR).join(DEFAULT_SECRETS_FILE));
     utils::write_secret_to_file(&secret_file, &secret_bytes, &secret_hash, &lock_txid)?;
     debug!(path = %secret_file.display(), "Secret written to file");
 
@@ -118,8 +107,8 @@ pub fn lock_bitcoin(args: LockBtcArgs) -> Result<LockBtcResult> {
 /// minted by revealing the secret.
 pub async fn commit_for_mint(args: CommitForMintArgs) -> Result<CommitResult> {
     match args.chain {
-        Chain::Ethereum => commit_for_mint_eth(args).await,
-        Chain::Solana => tokio::task::spawn_blocking(move || commit_for_mint_sol(args)).await?,
+        Chain::Ethereum => eth::commit_for_mint(args).await,
+        Chain::Solana => tokio::task::spawn_blocking(move || sol::commit_for_mint(args)).await?,
     }
 }
 
@@ -130,8 +119,8 @@ pub async fn commit_for_mint(args: CommitForMintArgs) -> Result<CommitResult> {
 /// claim the locked Bitcoin.
 pub async fn mint_with_secret(args: MintWithSecretArgs) -> Result<MintResult> {
     match args.chain {
-        Chain::Ethereum => mint_with_secret_eth(args).await,
-        Chain::Solana => tokio::task::spawn_blocking(move || mint_with_secret_sol(args)).await?,
+        Chain::Ethereum => eth::mint_with_secret(args).await,
+        Chain::Solana => tokio::task::spawn_blocking(move || sol::mint_with_secret(args)).await?,
     }
 }
 
@@ -231,8 +220,8 @@ pub fn claim_bitcoin(args: ClaimBtcArgs) -> Result<ClaimBtcResult> {
 /// can cancel. On Solana, only the seller can cancel.
 pub async fn cancel_commitment(args: CancelCommitArgs) -> Result<CancelResult> {
     match args.chain {
-        Chain::Ethereum => cancel_commitment_eth(args).await,
-        Chain::Solana => tokio::task::spawn_blocking(move || cancel_commitment_sol(args)).await?,
+        Chain::Ethereum => eth::cancel_commitment(args).await,
+        Chain::Solana => tokio::task::spawn_blocking(move || sol::cancel_commitment(args)).await?,
     }
 }
 
@@ -305,324 +294,5 @@ pub fn refund_bitcoin(args: RefundBtcArgs) -> Result<RefundBtcResult> {
         txid: refund_txid.to_string(),
         from_htlc,
         destination,
-    })
-}
-
-/// Commits an NFT for minting on Ethereum.
-async fn commit_for_mint_eth(args: CommitForMintArgs) -> Result<CommitResult> {
-    debug!("Executing NFT commitment for minting");
-
-    let rpc_url = args.eth_rpc.as_ref().unwrap();
-    let contract_addr = args.nft_contract.as_ref().unwrap();
-    let token_id = args.token_id;
-    let nft_price = args.nft_price;
-    let metadata_uri = args.metadata_uri;
-    let secret_hash = args.secret_hash;
-    let private_key = args.seller_eth_key.as_ref().unwrap();
-
-    let client = EthClient::new(rpc_url, private_key, *contract_addr)
-        .await
-        .context("Failed to initialize Ethereum client")?;
-
-    debug!(
-        seller_address = %client.get_address(),
-        "Connected to Ethereum as seller"
-    );
-
-    match client.get_commitment(U256::from(token_id)).await {
-        Ok(commitment) if commitment.is_active => {
-            return Err(anyhow!(
-                "Token {token_id} already has an active commitment from seller {:?}",
-                commitment.seller
-            ));
-        }
-        Ok(_) => {
-            debug!("No existing commitment found, proceeding");
-        }
-        Err(e) => {
-            debug!(error = %e, "Error checking existing commitment, proceeding anyway");
-        }
-    }
-
-    let tx_hash = client
-        .commit_for_mint(
-            H256(secret_hash),
-            U256::from(token_id),
-            U256::from(nft_price),
-            args.buyer_address,
-            metadata_uri.clone(),
-        )
-        .await
-        .context("Failed to commit NFT for minting")?;
-
-    debug!(
-        tx_hash = %tx_hash,
-        token_id = %token_id,
-        price_wei = %nft_price,
-        metadata_uri = %metadata_uri,
-        "NFT commitment transaction submitted"
-    );
-
-    Ok(CommitResult {
-        chain: args.chain.as_ref().to_string(),
-        tx_id: format!("{tx_hash:?}"),
-        token_id,
-        price: format!("{nft_price} wei"),
-        metadata_uri,
-    })
-}
-
-/// Commits an NFT for minting on Solana.
-fn commit_for_mint_sol(args: CommitForMintArgs) -> Result<CommitResult> {
-    debug!("Executing Solana NFT commitment for minting");
-
-    let program_id = args.program_id.as_ref().unwrap();
-    let keypair_path = args.seller_sol_keypair.as_ref().unwrap();
-    let rpc_url = args.sol_rpc.as_ref().unwrap();
-    let ws_url = args.sol_ws.as_ref().unwrap();
-
-    let token_id = args.token_id;
-    let secret_hash = args.secret_hash;
-    let name = args.name.as_ref().unwrap();
-    let symbol = args.symbol.as_ref().unwrap();
-    let metadata_uri = args.metadata_uri;
-    let nft_price = args.nft_price;
-
-    let payer = read_keypair_file(keypair_path).map_err(|e| anyhow!("{e}"))?;
-
-    let client = SolClient::new(payer, program_id, rpc_url, ws_url)
-        .context("Failed to initialize Solana client")?;
-
-    if !client.is_initialized() {
-        debug!("Program not initialized, attempting to initialize...");
-        let sig = client
-            .initialize()
-            .context("Failed to initialize Solana program")?;
-        debug!(signature = %sig, "Program initialized successfully");
-    }
-
-    match client.get_commitment(token_id) {
-        Ok(commitment) if !commitment.is_used => {
-            return Err(anyhow!(
-                "Token {token_id} already has an active commitment from {}",
-                commitment.seller
-            ));
-        }
-        Ok(_) => {
-            debug!("Commitment exists but is used, proceeding");
-        }
-        Err(e) => {
-            debug!(error = %e, "No existing commitment found, proceeding");
-        }
-    }
-
-    let sig = client
-        .commit_for_mint(
-            secret_hash,
-            token_id,
-            nft_price,
-            name.clone(),
-            symbol.clone(),
-            metadata_uri.clone(),
-        )
-        .context("Failed to commit NFT for minting on Solana")?;
-
-    debug!(
-        signature = %sig,
-        token_id = %token_id,
-        price_lamports = %nft_price,
-        name = %name,
-        symbol = %symbol,
-        metadata_uri = %metadata_uri,
-        "Solana NFT commitment transaction submitted"
-    );
-
-    Ok(CommitResult {
-        chain: args.chain.as_ref().to_string(),
-        tx_id: sig.to_string(),
-        token_id,
-        price: format!("{nft_price} lamports"),
-        metadata_uri,
-    })
-}
-
-/// Mints an NFT on Ethereum by revealing the secret.
-///
-/// Waits for the minimum commitment time to pass before attempting to mint.
-/// The revealed secret can then be used by the seller to claim the locked Bitcoin.
-async fn mint_with_secret_eth(args: MintWithSecretArgs) -> Result<MintResult> {
-    debug!("Executing NFT mint with secret reveal");
-
-    let rpc_url = args.eth_rpc.as_ref().unwrap();
-    let contract_addr = args.nft_contract.as_ref().unwrap();
-    let token_id = args.token_id;
-    let private_key = args.buyer_eth_key.as_ref().unwrap();
-    let secret = args.secret;
-
-    let client = EthClient::new(rpc_url, private_key, *contract_addr)
-        .await
-        .context("Failed to initialize Ethereum client")?;
-
-    let token_id_u256 = U256::from(token_id);
-    let secret_h256 = H256(secret);
-
-    if !client.can_mint_now(token_id_u256).await? {
-        debug!("Waiting for minimum commitment time to pass");
-        let wait_start = Instant::now();
-        loop {
-            if wait_start.elapsed() > MINT_AVAILABILITY_TIMEOUT {
-                return Err(anyhow!(
-                    "Timeout waiting for mint availability after {:?}",
-                    MINT_AVAILABILITY_TIMEOUT
-                ));
-            }
-
-            if client.can_mint_now(token_id_u256).await? {
-                debug!("Minimum commitment time passed, proceeding with mint");
-                break;
-            }
-
-            debug!("Mint not yet available, waiting {:?}", MINT_CHECK_INTERVAL);
-            sleep(MINT_CHECK_INTERVAL).await;
-        }
-    }
-
-    let tx_hash = client
-        .mint_with_secret(secret_h256, token_id_u256)
-        .await
-        .context("Failed to execute NFT mint transaction")?;
-
-    debug!(
-        tx_hash = %tx_hash,
-        secret_revealed = %hex::encode(secret),
-        "NFT minted successfully, secret revealed on Ethereum"
-    );
-
-    Ok(MintResult {
-        chain: args.chain.as_ref().to_string(),
-        tx_id: format!("{tx_hash:?}"),
-        token_id,
-        secret_revealed: hex::encode(secret),
-    })
-}
-
-/// Mints an NFT on Solana by revealing the secret.
-fn mint_with_secret_sol(args: MintWithSecretArgs) -> Result<MintResult> {
-    debug!("Executing Solana NFT mint with secret reveal");
-
-    let rpc_url = args.sol_rpc.as_ref().unwrap();
-    let ws_url = args.sol_ws.as_ref().unwrap();
-    let program_id = args.program_id.as_ref().unwrap();
-    let keypair_path = args.buyer_sol_keypair.as_ref().unwrap();
-    let payer = read_keypair_file(keypair_path).map_err(|e| anyhow!("{e}"))?;
-
-    let client = SolClient::new(payer, program_id, rpc_url, ws_url)
-        .context("Failed to initialize Solana client")?;
-
-    debug!(
-        buyer_address = %client.pubkey(),
-        "Connected to Solana as buyer"
-    );
-
-    let token_id = args.token_id;
-    let secret = args.secret;
-
-    let sig = client
-        .mint_with_secret(secret, token_id)
-        .context("Failed to execute Solana NFT mint transaction")?;
-
-    debug!(
-        signature = %sig,
-        secret_revealed = %hex::encode(secret),
-        token_id = %token_id,
-        "Solana NFT minted successfully, secret revealed"
-    );
-
-    Ok(MintResult {
-        chain: args.chain.as_ref().to_string(),
-        tx_id: sig.to_string(),
-        token_id,
-        secret_revealed: hex::encode(secret),
-    })
-}
-
-/// Cancels an NFT commitment on Ethereum.
-///
-/// Only the seller can cancel before the commitment timeout. After the timeout
-/// has passed, anyone can cancel the commitment to clean up expired state.
-async fn cancel_commitment_eth(args: CancelCommitArgs) -> Result<CancelResult> {
-    debug!("Executing Ethereum NFT commitment cancellation");
-
-    let rpc_url = args.eth_rpc.as_ref().unwrap();
-    let contract_addr = args.nft_contract.as_ref().unwrap();
-    let private_key = args.caller_eth_key.as_ref().unwrap();
-    let token_id = args.token_id;
-
-    let client = EthClient::new(rpc_url, private_key, *contract_addr)
-        .await
-        .context("Failed to initialize Ethereum client")?;
-
-    debug!(
-        caller_address = %client.get_address(),
-        token_id = %token_id,
-        "Connected to Ethereum, attempting to cancel commitment"
-    );
-
-    let tx_hash = client
-        .cancel_commitment(U256::from(token_id))
-        .await
-        .context("Failed to cancel commitment")?;
-
-    debug!(
-        tx_hash = %tx_hash,
-        token_id = %token_id,
-        "Commitment cancelled successfully"
-    );
-
-    Ok(CancelResult {
-        chain: args.chain.as_ref().to_string(),
-        tx_id: format!("{tx_hash:?}"),
-        token_id,
-    })
-}
-
-/// Cancels an NFT commitment on Solana.
-///
-/// Only the seller who created the commitment can cancel it. The commitment
-/// must not have been used (NFT not yet minted).
-fn cancel_commitment_sol(args: CancelCommitArgs) -> Result<CancelResult> {
-    debug!("Executing Solana NFT commitment cancellation");
-
-    let rpc_url = args.sol_rpc.as_ref().unwrap();
-    let ws_url = args.sol_ws.as_ref().unwrap();
-    let program_id = args.program_id.as_ref().unwrap();
-    let keypair_path = args.caller_sol_keypair.as_ref().unwrap();
-    let token_id = args.token_id;
-
-    let payer = read_keypair_file(keypair_path).map_err(|e| anyhow!("{e}"))?;
-
-    let client = SolClient::new(payer, program_id, rpc_url, ws_url)
-        .context("Failed to initialize Solana client")?;
-
-    debug!(
-        caller_address = %client.pubkey(),
-        token_id = %token_id,
-        "Connected to Solana, attempting to cancel commitment"
-    );
-
-    let sig = client
-        .cancel_commitment(token_id)
-        .context("Failed to cancel Solana commitment")?;
-
-    debug!(
-        signature = %sig,
-        token_id = %token_id,
-        "Solana commitment cancelled successfully"
-    );
-
-    Ok(CancelResult {
-        chain: args.chain.as_ref().to_string(),
-        tx_id: sig.to_string(),
-        token_id,
     })
 }
